Update new version

+++ modules/items/service/delete_item_by_id.go
@@ -1,9 +1,14 @@
 package itemservice
 
-import "context"
+import (
+	"context"
+	itemmodel "dev-coffee-api/modules/items/model"
+	"errors"
+)
 
 type DeleteItemByIdStorage interface {
 	DeleteItemById(ctx context.Context, id int) error
+	GetItemById(ctx context.Context, id int) (*itemmodel.Item, error)
 }
 
 type DeleteItemByIdService struct {
@@ -15,5 +20,13 @@ func NewDeleteItemByIdService(store DeleteItemByIdStorage) *DeleteItemByIdServic
 }
 
 func (s *DeleteItemByIdService) DeleteItemById(ctx context.Context, id int) error {
+	exist, err := s.store.GetItemById(ctx, id)
+	if err != nil {
+		return err
+	}
+
+	if exist == nil {
+		return errors.New("item not found")
+	}
 	return s.store.DeleteItemById(ctx, id)
 }
